+++ src/main/java/com/refit/domain/user/entity/type/LoginType.java
@@ -1,5 +1,9 @@
 package com.refit.domain.user.entity.type;
 
 public enum LoginType {
-    LOCAL, GOOGLE, KAKAO
+    LOCAL, GOOGLE, KAKAO;
+
+    public String toSignupServiceName() {
+        return name().toLowerCase() + "Signup";
+    }
 }
